data: Implement saving player data to flat file
The rest of the methods will be implemented in future commits. The file creation method is not in the utils class now because I don't intend for other classes to be creating files outside the main class.

There will be a different system in place for player inventories when they go to a new group for the first time. It will not involve reading from files as in the past.

Signed-off-by: Evan Maddock <maddock.evan@vivaldi.net>