Create the manager for adding and getting player profiles
This is a lot simpler than the old class because we don't need nearly as much logic due to using the Cache from Google; it takes care of invalidating cache entries for us. We also no longer need the old recurring task to run to take care of saving/invalidation logic.

Caching has also been moved to the profile manager, since it makes way more sense to cache on creation than to cache on load. This also removes the need to manually invalidate cache entries when an updated profile is saved.

Signed-off-by: Gnat008 <maddock.evan@gmail.com>