Add a constructor for PlayerProfiles so that they can be built directly from values
This means that we can construct a profile back from Json, and eventually from a MySQL ResultSet.

This also means that we no longer just set things from raw Json; we have an actual profile object to work with. That way deserializing is actually deserializing, not actually applying. These things should be separate.

Signed-off-by: Evan Maddock <maddock.evan@vivaldi.net>