Replace testing framework and redo group commands tests
Switching out the Mockito framework allows us to a) use one made specifically for Kotlin and b) write nicer and cleaner tests without resorting to various hacks (mainly with mocking).

Eventually all tests should be redone in the new framework, but it's not vital and they all still work fine.

Signed-off-by: Evan Maddock <maddock.evan@vivaldi.net>